+++ spec/unit/language/syntax/to_ast/test_task_def.rb
@@ -1,28 +1,42 @@
 require 'spec_helper'
 module Gisele::Language::Syntax
-  describe ExplicitTaskDef, "to_ast" do
+  describe TaskDef, "to_ast" do
 
     it 'parses an explicit definition as expected' do
-      expr     = "task Task1 fluent diagKnown {}, {} refinement Task2 end end"
+      expr = <<-EXPR.strip
+        task Task1
+          fluent diagKnown {}, {}
+          Task2
+        end
+      EXPR
       expected = \
         [:task_def, "Task1",
-          [:task_signature,
-            [:fluent, "diagKnown", [:event_set], [:event_set], nil]],
-          [:task_refinement,
-            [:task_call_st, "Task2"]]]
+          [:fluent, "diagKnown", [:event_set], [:event_set], nil],
+          [:task_call_st, "Task2"]]
       ast(expr, :task_def).should eq(expected)
     end
 
-  end
-  describe ImplicitTaskDef, "to_ast" do
-
-    it 'parses an implicit definition as expected' do
-      expr     = "task Task1 Task2 end"
+    it 'uses :nop if no statement' do
+      expr = <<-EXPR.strip
+        task Task1
+          fluent diagKnown {}, {}
+        end
+      EXPR
       expected = \
         [:task_def, "Task1",
-          [:task_signature],
-          [:task_refinement,
-            [:task_call_st, "Task2"]]]
+          [:fluent, "diagKnown", [:event_set], [:event_set], nil],
+          [:nop]]
+      ast(expr, :task_def).should eq(expected)
+    end
+
+    it 'uses :nop when empty' do
+      expr = <<-EXPR.strip
+        task Task1
+        end
+      EXPR
+      expected = \
+        [:task_def, "Task1",
+          [:nop]]
       ast(expr, :task_def).should eq(expected)
     end